Galen's suggested changes to Rev 1.0
I have changed a number of things in your design, one of which is major and the
rest are minor. The major change is that I have added a guard trace to your
design, in accordance with the data sheet for the OPA129. If you are going to
use an ultra-low input bias current op amp, then you should make efforts to
actually achieve the performance that it touts. The other changes are mostly
positioning to accommodate the guard trace, changes to the silk screen on your
board, and simplification of the routing of some traces.

To achieve very low input bias, you need to make sure that there is no leakage
current to the electrode net from the rest of the PCB. There is some parasitic
resistance between neighboring traces in the PCB through the insulator and the
solder-mask. By surrounding the electrode with a copper trace that is driven to
be the same voltage as the electrode, this eliminates parasitic resistance, and
also reduces the stray capacitance of the electrode. I am mainly just following
the recommendations of the data sheet for the OPA129, so for more details see the
data sheet. I moved the bridge resistor and compensation capacitor closer to the
electrode to minimize the length of the electrode trace on your board, since
it's a high impedance trace that should be minimized and protected to avoid
noise pickup. I also placed the cap and res so that the guard trace runs between
the pads. To be honest, with these in place the point of using the OPA129 is
lost on me, since you'll have more parasitic current through these than through
any op amp. However, if you do not place them (which you probably will not need
to), then you will have very little input bias current. Running the trace
through the pads is manufacturable at PCBWay (0.1mm minimum), but may cost more.
I think the difference in cost will be totally worth it. We're not that price
sensitive. I also added a break in the solder mask for the guard traces, once
again to minimize parasitic current leaking into your electrode.

In a few cases I simplified your routing to eliminate right angles, which makes
it easier to manufacture and is better for reflections etc for high speed
signals. In the case of your board, none of this really matters, but it also
looks nice!

I moved the silkscreen around a bit to make it more visible, and moved the
'glass' silk screen to the back of the board where you will actually see it.

I also added the name of the PCB and the revision. I think adding the revision
is critical since these PCBs invariably go through a number of iterations, and
trying to figure out which PCB you're looking at can get difficult. I would
recommend tagging the git repo with the revision number once you send this to
the fab as well.