Remove screenshots from repo
These screen shots are 'derived' from the design files and should not be in the
repository itself. For example, are they currently up to date, or do they
reflect the design as it was a few commits ago? It's hard to tell, and therefore
I think these sorts of files do more harm than good.